Add files via upload
The video created here is not intended for harming any individual. It's just a study of how DeepFakes are becoming increasingly used and through its simplicity can be tranformed from one form to another. The files that I have uploaded in this section portrays the simplicity by which DeepFakes could be made!!